+++ core/backend/src/test/ecdb/ECDbTestHelper.ts
@@ -8,24 +8,23 @@ import { ECDb } from "../../ECDb";
 import { IModelJsFs } from "../../IModelJsFs";
 
 export class ECDbTestHelper {
-
-  public static createECDb(outDir: string, fileName: string, schemaXml?: string): ECDb {
-    if (!IModelJsFs.existsSync(outDir))
-      IModelJsFs.mkdirSync(outDir);
+  public static createECDb(
+    outDir: string,
+    fileName: string,
+    schemaXml?: string
+  ): ECDb {
+    if (!IModelJsFs.existsSync(outDir)) IModelJsFs.mkdirSync(outDir);
 
     const outPath = path.join(outDir, fileName);
-    if (IModelJsFs.existsSync(outPath))
-      IModelJsFs.unlinkSync(outPath);
+    if (IModelJsFs.existsSync(outPath)) IModelJsFs.unlinkSync(outPath);
 
     const ecdb = new ECDb();
     ecdb.createDb(outPath);
 
-    if (!schemaXml)
-      return ecdb;
+    if (!schemaXml) return ecdb;
 
     const schemaPath = path.join(outDir, `${Guid.createValue()}.ecschema.xml`);
-    if (IModelJsFs.existsSync(schemaPath))
-      IModelJsFs.unlinkSync(schemaPath);
+    if (IModelJsFs.existsSync(schemaPath)) IModelJsFs.unlinkSync(schemaPath);
 
     IModelJsFs.writeFileSync(schemaPath, schemaXml);
